fix error and connect with server

+++ src/Menu_creation_partie.cpp
@@ -187,8 +187,8 @@ info Menu_creation_partie::run(info information)
         {
 
           information.client->createRoom(information.username);
-          information.client->setTimeRound(nombre3);//durée par tours
-          information.client->set_nrb_lombrics(nombre2);//nombre lombrics
+          information.client->setTimeRound(static_cast<uint32_t>(nombre3));//durée par tours
+          information.client->set_nrb_lombrics(static_cast<uint32_t>(nombre2));//nombre lombrics
           //information.client->set_nbr_equipes(nombre1);
           information.id=28;
           break;
